+++ src/main/java/me/crylonz/DeadChest.java
@@ -2,11 +2,12 @@ package me.crylonz;
 
 import me.crylonz.commands.DCCommandExecutor;
 import me.crylonz.commands.TabCompletion;
+import me.crylonz.utils.ConfigKey;
+import me.crylonz.utils.DeadChestConfig;
+import me.crylonz.utils.DeadChestUpdater;
 import org.bstats.bukkit.Metrics;
 import org.bukkit.Material;
 import org.bukkit.World;
-import org.bukkit.configuration.file.FileConfiguration;
-import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.configuration.serialization.ConfigurationSerialization;
 import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.PluginDescriptionFile;
@@ -31,30 +32,10 @@ public class DeadChest extends JavaPlugin {
     public static Localization local;
     public static Plugin plugin;
 
-    // Config
-    public static boolean isIndestructible = true;
-    public static boolean OnlyOwnerCanOpenDeadChest = true;
-    public static int chestDuration = 600;
-    public static int maxDeadChestPerPlayer = 5;
-    public static boolean logDeadChestOnConsole = false;
-    public static boolean requirePermissionToGenerate = false;
-    public static boolean requirePermissionToGetChest = false;
-    public static boolean permissionRequiredToListOwn = false;
-    public static boolean autoCleanUpOnStart = false;
-    public static boolean generateDeadChestInCreative = true;
-    public static boolean displayDeadChestPositionOnDeath = true;
-    public static int dropMode = 1;
-    public static int dropBlock = 1;
-    public static ArrayList<String> excludedWorlds = new ArrayList<>();
-    public static ArrayList<String> excludedItems = new ArrayList<>();
-    public static boolean itemsDroppedAfterTimeOut = false;
-    public static boolean enableWorldGuardDetection = false;
-    public static boolean generateOnLava = true;
-    public static boolean generateOnWater = true;
-    public static boolean generateOnRails = true;
-    public static boolean generateInMinecart = true;
     public static boolean bstats = true;
     public static boolean isChangesNeedToBeSave = false;
+
+    public static DeadChestConfig config;
 
     static {
         ConfigurationSerialization.registerClass(ChestData.class, "ChestData");
@@ -68,43 +49,10 @@ public class DeadChest extends JavaPlugin {
         super(loader, description, dataFolder, file);
     }
 
-    public static void handleEvent() {
-        if (chestData != null && !chestData.isEmpty()) {
-
-            Date now = new Date();
-            Iterator<ChestData> chestDataIt = chestData.iterator();
-
-            while (chestDataIt.hasNext()) {
-                ChestData chestData = chestDataIt.next();
-                World world = chestData.getChestLocation().getWorld();
-
-                if (world != null) {
-                    updateTimer(chestData, now);
-
-                    if (handleExpirateDeadChest(chestData, chestDataIt, now)) {
-                        isChangesNeedToBeSave = true;
-                        generateLog("Deadchest of [" + chestData.getPlayerName() + "] has expired in " + Objects.requireNonNull(chestData.getChestLocation().getWorld()).getName());
-                    } else {
-                        if (chestData.isChunkLoaded()) {
-                            isChangesNeedToBeSave = replaceDeadChestIfItDeseapears(chestData);
-                        }
-                    }
-                }
-            }
-        }
-
-        if (isChangesNeedToBeSave) {
-            fileManager.saveModification();
-            isChangesNeedToBeSave = false;
-        }
-    }
-
     public void onEnable() {
 
-        plugin = this;
         PluginManager pm = getServer().getPluginManager();
-        pm.registerEvents(new DeadChestListener(), this);
-        fileManager = new FileManager(this);
+        pm.registerEvents(new DeadChestListener(this), this);
 
         // Wich block can be used as grave ?
         graveBlocks.add(Material.CHEST);
@@ -113,34 +61,42 @@ public class DeadChest extends JavaPlugin {
         graveBlocks.add(Material.BARREL);
         graveBlocks.add(Material.SHULKER_BOX);
 
+
+        Objects.requireNonNull(this.getCommand("dc"), "Command dc not found")
+                .setExecutor(new DCCommandExecutor(this));
+
         Objects.requireNonNull(getCommand("dc")).setTabCompleter(new TabCompletion());
 
         if (bstats) {
             Metrics metrics = new Metrics(this, 11385);
         }
 
-        chestData = new ArrayList<>();
-        local = new Localization();
-
-        Objects.requireNonNull(this.getCommand("dc"), "Command dc not found")
-                .setExecutor(new DCCommandExecutor(this));
-
-        initializeConfig();
-
-        if (autoCleanUpOnStart) {
-            cleanAllDeadChests();
-        }
         launchRepeatingTask();
     }
 
     @Override
     public void onLoad() {
 
-        // to be sure that data is loaded when call onLoad()
-        enableWorldGuardDetection = (boolean) getConfig().get("EnableWorldGuardDetection");
+        config = new DeadChestConfig(this);
+        plugin = this;
+        fileManager = new FileManager(this);
+
+        chestData = new ArrayList<>();
+        local = new Localization();
 
 
-        if (enableWorldGuardDetection) {
+        registerConfig();
+        initializeConfig();
+
+        if (config.getBoolean(ConfigKey.AUTO_UPDATE)) {
+            DeadChestUpdater updater = new DeadChestUpdater(this, 322882, this.getFile(), DeadChestUpdater.UpdateType.DEFAULT, true);
+        }
+
+        if (config.getBoolean(ConfigKey.AUTO_CLEANUP_ON_START)) {
+            cleanAllDeadChests();
+        }
+
+        if (config.getBoolean(ConfigKey.WORLD_GUARD_DETECTION)) {
 
             try {
                 wgsdc = new WorldGuardSoftDependenciesChecker();
@@ -163,56 +119,49 @@ public class DeadChest extends JavaPlugin {
         }
     }
 
+    public void registerConfig() {
+        config.register(ConfigKey.AUTO_UPDATE.toString(), true);
+        config.register(ConfigKey.INDESTRUCTIBLE_CHEST.toString(), true);
+        config.register(ConfigKey.ONLY_OWNER_CAN_OPEN_CHEST.toString(), true);
+        config.register(ConfigKey.DEADCHEST_DURATION.toString(), 300);
+        config.register(ConfigKey.MAX_DEAD_CHEST_PER_PLAYER.toString(), 15);
+        config.register(ConfigKey.LOG_DEADCHEST_ON_CONSOLE.toString(), false);
+        config.register(ConfigKey.REQUIRE_PERMISSION_TO_GENERATE.toString(), false);
+        config.register(ConfigKey.REQUIRE_PERMISSION_TO_GET_CHEST.toString(), false);
+        config.register(ConfigKey.REQUIRE_PERMISSION_TO_LIST_OWN.toString(), false);
+        config.register(ConfigKey.AUTO_CLEANUP_ON_START.toString(), false);
+        config.register(ConfigKey.GENERATE_DEADCHEST_IN_CREATIVE.toString(), true);
+        config.register(ConfigKey.DISPLAY_POSITION_ON_DEATH.toString(), true);
+        config.register(ConfigKey.ITEMS_DROPPED_AFTER_TIMEOUT.toString(), false);
+        config.register(ConfigKey.WORLD_GUARD_DETECTION.toString(), false);
+        config.register(ConfigKey.DROP_MODE.toString(), 1);
+        config.register(ConfigKey.DROP_BLOCK.toString(), 1);
+        config.register(ConfigKey.GENERATE_ON_LAVA.toString(), true);
+        config.register(ConfigKey.GENERATE_ON_WATER.toString(), true);
+        config.register(ConfigKey.GENERATE_ON_RAILS.toString(), true);
+        config.register(ConfigKey.GENERATE_IN_MINECART.toString(), true);
+        config.register(ConfigKey.EXCLUDED_WORLDS.toString(), Collections.emptyList());
+        config.register(ConfigKey.EXCLUDED_ITEMS.toString(), Collections.emptyList());
+    }
+
     private void initializeConfig() {
 
         // plugin config file
         if (!fileManager.getConfigFile().exists()) {
             saveDefaultConfig();
-
         } else {
-
-            autoUpdateConfigFile();
-
-            @SuppressWarnings("unchecked")
-            ArrayList<ChestData> tmp = (ArrayList<ChestData>) fileManager.getChestDataConfig().get("chestData");
-
-            @SuppressWarnings("unchecked")
-            ArrayList<String> tmpExludedWorld = (ArrayList<String>) getConfig().get("ExcludedWorld");
-            ArrayList<String> tmpExcludedItems = (ArrayList<String>) getConfig().get("ExcludedItems");
-
-            if (tmp != null)
-                chestData = tmp;
-
-            if (tmpExludedWorld != null)
-                excludedWorlds = tmpExludedWorld;
-
-            if (tmpExcludedItems != null)
-                excludedItems = tmpExcludedItems;
-
-            isIndestructible = getConfig().getBoolean("IndestuctibleChest");
-            OnlyOwnerCanOpenDeadChest = getConfig().getBoolean("OnlyOwnerCanOpenDeadChest");
-            chestDuration = getConfig().getInt("DeadChestDuration");
-            maxDeadChestPerPlayer = (int) getConfig().get("maxDeadChestPerPlayer");
-            logDeadChestOnConsole = (boolean) getConfig().get("logDeadChestOnConsole");
-            requirePermissionToGenerate = (boolean) getConfig().get("RequirePermissionToGenerate");
-            requirePermissionToGetChest = (boolean) getConfig().get("RequirePermissionToGetChest");
-            permissionRequiredToListOwn = (boolean) getConfig().get("RequirePermissionToListOwn");
-            autoCleanUpOnStart = (boolean) getConfig().get("AutoCleanupOnStart");
-            generateDeadChestInCreative = (boolean) getConfig().get("GenerateDeadChestInCreative");
-            displayDeadChestPositionOnDeath = (boolean) getConfig().get("DisplayDeadChestPositionOnDeath");
-            itemsDroppedAfterTimeOut = (boolean) getConfig().get("ItemsDroppedAfterTimeOut");
-            enableWorldGuardDetection = (boolean) getConfig().get("EnableWorldGuardDetection");
-            dropMode = (int) getConfig().get("DropMode");
-            dropBlock = (int) getConfig().get("DropBlock");
-            generateOnLava = (boolean) getConfig().get("GenerateOnLava");
-            generateOnWater = (boolean) getConfig().get("GenerateOnWater");
-            generateOnRails = (boolean) getConfig().get("GenerateOnRails");
-            generateInMinecart = (boolean) getConfig().get("GenerateInMinecart");
+            config.updateConfig();
         }
 
         // database (chestData.yml)
         if (!fileManager.getChestDataFile().exists()) {
             fileManager.saveChestDataConfig();
+        } else {
+            @SuppressWarnings("unchecked")
+            ArrayList<ChestData> tmp = (ArrayList<ChestData>) fileManager.getChestDataConfig().get("chestData");
+
+            if (tmp != null)
+                chestData = tmp;
         }
 
         // locale file for translation
@@ -254,7 +203,7 @@ public class DeadChest extends JavaPlugin {
             );
         } else {
             // if file exist
-            // we verify if the file have all traduction
+            // we verify if the file have all translation
             // and add missing if needed
 
             Map<String, Object> localTmp =
@@ -271,52 +220,42 @@ public class DeadChest extends JavaPlugin {
         fileManager.saveLocalizationConfig();
     }
 
+    public static void handleEvent() {
+        if (chestData != null && !chestData.isEmpty()) {
+
+            Date now = new Date();
+            Iterator<ChestData> chestDataIt = chestData.iterator();
+
+            while (chestDataIt.hasNext()) {
+                ChestData chestData = chestDataIt.next();
+                World world = chestData.getChestLocation().getWorld();
+
+                if (world != null) {
+                    updateTimer(chestData, now);
+
+                    if (handleExpirateDeadChest(chestData, chestDataIt, now)) {
+                        isChangesNeedToBeSave = true;
+                        generateLog("Deadchest of [" + chestData.getPlayerName() + "] has expired in " + Objects.requireNonNull(chestData.getChestLocation().getWorld()).getName());
+                    } else {
+                        if (chestData.isChunkLoaded()) {
+                            isChangesNeedToBeSave = replaceDeadChestIfItDeseapears(chestData);
+                        }
+                    }
+                }
+            }
+        }
+
+        if (isChangesNeedToBeSave) {
+            fileManager.saveModification();
+            isChangesNeedToBeSave = false;
+        }
+    }
+
     private void launchRepeatingTask() {
         getServer().getScheduler().scheduleSyncRepeatingTask(this, DeadChest::handleEvent, 20, 20);
     }
 
-    // Add missing parameters on config.yml
-    public void autoUpdateConfigFile() {
-        reloadConfig();
-        ArrayList<String> allConfigPath = new ArrayList<>();
-        allConfigPath.add("IndestuctibleChest");
-        allConfigPath.add("OnlyOwnerCanOpenDeadChest");
-        allConfigPath.add("DeadChestDuration");
-        allConfigPath.add("maxDeadChestPerPlayer");
-        allConfigPath.add("logDeadChestOnConsole");
-        allConfigPath.add("RequirePermissionToGenerate");
-        allConfigPath.add("RequirePermissionToGetChest");
-        allConfigPath.add("RequirePermissionToListOwn");
-        allConfigPath.add("AutoCleanupOnStart");
-        allConfigPath.add("GenerateDeadChestInCreative");
-        allConfigPath.add("DisplayDeadChestPositionOnDeath");
-        allConfigPath.add("DropMode");
-        allConfigPath.add("DropBlock");
-        allConfigPath.add("ExcludedWorld");
-        allConfigPath.add("ExcludedItems");
-        allConfigPath.add("ItemsDroppedAfterTimeOut");
-        allConfigPath.add("EnableWorldGuardDetection");
-        allConfigPath.add("GenerateOnLava");
-        allConfigPath.add("GenerateOnWater");
-        allConfigPath.add("GenerateOnRails");
-        allConfigPath.add("GenerateInMinecart");
-
-        for (String path : allConfigPath) {
-            FileConfiguration configuration = YamlConfiguration.loadConfiguration(fileManager.getConfigFile());
-            if (configuration.get(path) == null) {
-                File file = new File(getDataFolder().getAbsolutePath() + File.separator + "config.yml");
-                File oldFile = new File(getDataFolder().getAbsolutePath() + File.separator + "config.old.yml");
-                file.renameTo(oldFile);
-                saveDefaultConfig();
-                log.warning("---------------------------------------");
-                log.warning("");
-                log.warning("[DeadChest] Configuration update detected !");
-                log.warning("[DeadChest] Config back to default : Don't forget to update your config.yml");
-                log.warning("[DeadChest] Your old config is in config.old.yml");
-                log.warning("");
-                log.warning("---------------------------------------");
-                break;
-            }
-        }
+    public DeadChestConfig getDataConfig() {
+        return config;
     }
 }