+++ src/main/java/me/crylonz/DeadChestManager.java
@@ -1,5 +1,6 @@
 package me.crylonz;
 
+import me.crylonz.utils.ConfigKey;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
@@ -12,7 +13,10 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.metadata.FixedMetadataValue;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Date;
+import java.util.Iterator;
 
 import static me.crylonz.DeadChest.*;
 import static me.crylonz.Utils.computeChestType;
@@ -133,20 +137,20 @@ public class DeadChestManager {
     }
 
     public static boolean handleExpirateDeadChest(ChestData chestData, Iterator<ChestData> chestDataIt, Date date) {
-        if (chestData.getChestDate().getTime() + chestDuration * 1000L < date.getTime() && !chestData.isInfinity()
-                && chestDuration != 0) {
+        if (chestData.getChestDate().getTime() + config.getInt(ConfigKey.DEADCHEST_DURATION) * 1000L < date.getTime() && !chestData.isInfinity()
+                && config.getInt(ConfigKey.DEADCHEST_DURATION) != 0) {
 
             Location loc = chestData.getChestLocation();
 
             if (loc.getWorld() != null) {
                 loc.getWorld().getBlockAt(loc).setType(Material.AIR);
-                if (itemsDroppedAfterTimeOut) {
+                if (config.getBoolean(ConfigKey.ITEMS_DROPPED_AFTER_TIMEOUT)) {
                     for (ItemStack itemStack : chestData.getInventory()) {
                         if (itemStack != null) {
                             loc.getWorld().dropItemNaturally(loc, itemStack);
                         }
                     }
-                     chestData.cleanInventory();
+                    chestData.cleanInventory();
                 }
             }
             if (chestData.removeArmorStand()) {
@@ -170,12 +174,12 @@ public class DeadChestManager {
                         reloadMetaData();
                     }
                     if (entity.getMetadata("deadchest").size() > 0 && entity.getMetadata("deadchest").get(0).asBoolean()) {
-                        long diff = date.getTime() - (chestData.getChestDate().getTime() + chestDuration * 1000L);
+                        long diff = date.getTime() - (chestData.getChestDate().getTime() + config.getInt(ConfigKey.DEADCHEST_DURATION) * 1000L);
                         long diffSeconds = Math.abs(diff / 1000 % 60);
                         long diffMinutes = Math.abs(diff / (60 * 1000) % 60);
                         long diffHours = Math.abs(diff / (60 * 60 * 1000));
 
-                        if (!chestData.isInfinity() && chestDuration != 0) {
+                        if (!chestData.isInfinity() && config.getInt(ConfigKey.DEADCHEST_DURATION) != 0) {
                             entity.setCustomName(local.replaceTimer(local.get("holo_timer"), diffHours, diffMinutes, diffSeconds));
                         } else {
                             entity.setCustomName(local.get("loc_infinityChest"));